Refactor to match latest AppRouter
