+++ src/containers/App/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import AppRouter from './routers/AppRouter';
+import AppRouter from '../../routers/AppRouter';
 import './App.css';
 
 class App extends Component {
