+++ src/routers/AppRouter.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {BrowserRouter, Route, Switch} from 'react-router-dom'
 import Header from '../components/Header';
 import Dashboard from '../components/Dashboard';
-import '../App.css';
+
 const AppRouter = () => (
   <BrowserRouter>
   <div>
